requirements.txt
Django==5.2.8
django-mdeditor==0.1.20
python-dotenv==1.2.1